LC - Match only initial conditional block part in `extract_tag_names`
We were trying to match whole conditionaal block tags in
`extract_tag_names` and that was causing us to miss some broken tags in
`unusable_tags` method.

Also, we weren't matching some conditional tags because the '^' part
wasn't being escaped when creating the the regex. Now the generated
regexes are being properly escaped.