LC - Fix tests according to the latest changes in `Document` logic
Despite the simple test expectations change, it just means theat we are
not considering closing conditional tags and invalid opening conditional
tags (that were being considered in the old test numbers). So, with the
latest commits, the methods `usable_tags` and `usable_tag_names` (and
others) just became thriftier than before and that won't compromise the
expected features for this gem in general.