LC - Remove usage of `text_nodes_containing_only_starting_conditionals`
We were using the `text_nodes_containing_only_starting_conditionals`
method in `unusable_tags` just because we weren't including in
`usable_tags` the tags that were coming from start conditional block
tags that the only content inside a paragraph. Since this issue is fixed
in the `usable_tags` method itself, this workaround is being removed.